Feat: Add basic YAML-1.1 support for booleans
Pre YAML-1.2, words such as `yes`, `no`, `on` and `off` were regarded
as boolean values. This move brings the language more in line with
other languages, but can cause issues with integrations where older
specs of YAML are used, where we have generated the documents using a
newer spec.

Specifically looking at Kubernetes (e.g. 1.25) this issue can be seen when
generating manifests as `no` strings in argument lists will be output
as `no` rather than `'no'`.

This change brings back the ability to to output more in line with the
pre YAML-1.2 spec.